+++ demisto_sdk/commands/common/hook_validations/readme.py
@@ -207,7 +207,9 @@ class ReadMeValidator(BaseValidator):
                 self.verify_template_not_in_readme(),
                 self.verify_copyright_section_in_readme_content(),
                 # self.has_no_markdown_lint_errors(),
-            ]
+                self.validate_no_disallowed_terms_in_customer_facing_docs(file_content=self.readme_content,
+                                                                          file_path=self.file_path_str),
+                ]
         )
 
     def mdx_verify_server(self) -> bool: